+++ packages/react/src/types.ts
@@ -6,7 +6,9 @@ import type {
   UploadThingError,
 } from "@uploadthing/shared";
 import type {
+  AnyFileRoute,
   ClientUploadedFileData,
+  EndpointArg,
   FileRouter,
   inferEndpointInput,
   inferEndpointOutput,
@@ -27,9 +29,8 @@ export interface GenerateTypedHelpersOptions {
 }
 
 export type UseUploadthingProps<
-  TRouter extends FileRouter,
-  TEndpoint extends keyof TRouter,
-  TServerOutput = inferEndpointOutput<TRouter[TEndpoint]>,
+  TFileRoute extends AnyFileRoute,
+  TServerOutput = inferEndpointOutput<TFileRoute>,
 > = {
   /**
    * Called when the upload is submitted and the server is about to be queried for presigned URLs
@@ -73,7 +74,7 @@ export type UseUploadthingProps<
    * Called if the upload fails
    */
   onUploadError?:
-    | ((e: UploadThingError<inferErrorShape<TRouter>>) => MaybePromise<void>)
+    | ((e: UploadThingError<inferErrorShape<TFileRoute>>) => MaybePromise<void>)
     | undefined;
   /**
    * Set custom headers that'll get sent with requests
@@ -93,7 +94,7 @@ export type UploadthingComponentProps<
   TRouter extends FileRouter,
   TEndpoint extends keyof TRouter,
 > = Omit<
-  UseUploadthingProps<TRouter, TEndpoint>,
+  UseUploadthingProps<TRouter[TEndpoint]>,
   /**
    * Signal is omitted, component has its own AbortController
    * If you need to control the interruption with more granularity,
@@ -110,7 +111,7 @@ export type UploadthingComponentProps<
   /**
    * The endpoint from your FileRouter to use for the upload
    */
-  endpoint: TEndpoint;
+  endpoint: EndpointArg<TRouter, TEndpoint>;
   /**
    * URL to the UploadThing API endpoint
    * @example URL { /api/uploadthing }