+++ packages/shared/src/utils.ts
@@ -17,6 +17,7 @@ import type {
   FileRouterInputConfig,
   FileRouterInputKey,
   FileSize,
+  Json,
   ResponseEsque,
   RouteConfig,
   Time,
@@ -167,17 +168,8 @@ export const bytesToFileSize = (bytes: number) => {
 };
 
 export async function safeParseJSON<T>(
-  input: string | ResponseEsque | Request,
+  input: ResponseEsque,
 ): Promise<T | Error> {
-  if (typeof input === "string") {
-    try {
-      return JSON.parse(input) as T;
-    } catch (err) {
-      console.error(`Error parsing JSON, got '${input}'`);
-      return new Error(`Error parsing JSON, got '${input}'`);
-    }
-  }
-
   const text = await input.text();
   try {
     return JSON.parse(text ?? "null") as T;
@@ -192,15 +184,6 @@ export function objectKeys<T extends Record<string, unknown>>(
   obj: T,
 ): (keyof T)[] {
   return Object.keys(obj) as (keyof T)[];
-}
-
-/** checks if obj is a valid, non-null object */
-export function isObject(obj: unknown): obj is Record<string, unknown> {
-  return typeof obj === "object" && obj !== null && !Array.isArray(obj);
-}
-
-export function asArray<T>(val: T | T[]): T[] {
-  return Array.isArray(val) ? val : [val];
 }
 
 export function filterDefinedObjectValues<T>(
@@ -351,3 +334,20 @@ export const safeNumberReplacer = (_: string, value: unknown) => {
   if (value === -Infinity) return Number.MIN_SAFE_INTEGER;
   if (Number.isNaN(value)) return 0;
 };
+
+export function noop() {
+  // noop
+}
+
+export function createIdentityProxy<TObj extends Record<string, unknown>>() {
+  return new Proxy(noop, {
+    get: (_, prop) => prop,
+  }) as unknown as TObj;
+}
+
+export function unwrap<T extends Json | PropertyKey, Param extends unknown[]>(
+  x: T | ((...args: Param) => T),
+  ...args: Param
+) {
+  return typeof x === "function" ? x(...args) : x;
+}