+++ packages/uploadthing/src/internal/route-config.ts
@@ -1,6 +1,6 @@
-import type * as S from "@effect/schema/Schema";
 import * as Data from "effect/Data";
 import * as Effect from "effect/Effect";
+import type * as S from "effect/Schema";
 
 import type {
   ExpandedRouteConfig,
@@ -20,8 +20,8 @@ import {
   UploadThingError,
 } from "@uploadthing/shared";
 
+import type { FileRouter } from "../types";
 import type { UploadActionPayload } from "./shared-schemas";
-import type { FileRouter } from "./types";
 
 class FileSizeMismatch extends Data.Error<{
   reason: string;
@@ -118,8 +118,8 @@ export const extractRouterConfig = <TRouter extends FileRouter>(
   router: TRouter,
 ) =>
   Effect.forEach(objectKeys(router), (slug) =>
-    Effect.map(
-      fillInputRouteConfig(router[slug]._def.routerConfig),
-      (config) => ({ slug, config }),
-    ),
+    Effect.map(fillInputRouteConfig(router[slug].routerConfig), (config) => ({
+      slug,
+      config,
+    })),
   );