+++ packages/uploadthing/src/next.ts
@@ -4,18 +4,22 @@ import * as Effect from "effect/Effect";
 import type { Json } from "@uploadthing/shared";
 
 import { makeAdapterHandler } from "./internal/handler";
-import type { FileRouter, RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
+import type { FileRouter, RouteHandlerOptions } from "./types";
 
 export type { FileRouter };
 export { UTFiles } from "./internal/types";
 
-type MiddlewareArgs = { req: NextRequest; res: undefined; event: undefined };
+type AdapterArgs = {
+  req: NextRequest;
+  res: undefined;
+  event: undefined;
+};
 
 export const createUploadthing = <TErrorShape extends Json>(
   opts?: CreateBuilderOptions<TErrorShape>,
-) => createBuilder<MiddlewareArgs, TErrorShape>(opts);
+) => createBuilder<AdapterArgs, TErrorShape>(opts);
 
 export const createRouteHandler = <TRouter extends FileRouter>(
   opts: RouteHandlerOptions<TRouter>,