+++ packages/uploadthing/src/server.ts
@@ -5,20 +5,24 @@ import { UploadThingError } from "@uploadthing/shared";
 
 import { makeAdapterHandler } from "./internal/handler";
 import { extractRouterConfig as extractEffect } from "./internal/route-config";
-import type { FileRouter, RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
+import type { FileRouter, RouteHandlerOptions } from "./types";
 
 export { UTFiles } from "./internal/types";
 export { UTApi } from "./sdk";
 export { UTFile } from "./sdk/ut-file";
 export { UploadThingError, type FileRouter };
 
-type MiddlewareArgs = { req: Request; res: undefined; event: undefined };
+type AdapterArgs = {
+  req: Request;
+  res: undefined;
+  event: undefined;
+};
 
 export const createUploadthing = <TErrorShape extends Json>(
   opts?: CreateBuilderOptions<TErrorShape>,
-) => createBuilder<MiddlewareArgs, TErrorShape>(opts);
+) => createBuilder<AdapterArgs, TErrorShape>(opts);
 
 export const createRouteHandler = <TRouter extends FileRouter>(
   opts: RouteHandlerOptions<TRouter>,